requirements.txt
pandas
requests
transformers
torch
nltk
python-dotenv
matplotlib